Remove old classes with tests
