+++ src/Keboola/DbExtractor/Extractor/BaseExtractor.php
@@ -26,6 +26,8 @@ abstract class BaseExtractor extends BaseComponent
 {
     public const DEFAULT_MAX_TRIES = 5;
 
+    public const DATATYPE_KEYS = ['type', 'length', 'nullable', 'default', 'format'];
+
     /** @var array */
     protected $dbParameters;
 
@@ -294,13 +296,12 @@ abstract class BaseExtractor extends BaseComponent
 
     private function getColumnMetadata(array $column): array
     {
-        $datatypeKeys = ['type', 'length', 'nullable', 'default', 'format'];
         $datatype = new GenericStorage(
             $column['type'],
-            array_intersect_key($column, array_flip($datatypeKeys))
+            array_intersect_key($column, array_flip(self::DATATYPE_KEYS))
         );
         $columnMetadata = $datatype->toMetadata();
-        $nonDatatypeKeys = array_diff_key($column, array_flip($datatypeKeys));
+        $nonDatatypeKeys = array_diff_key($column, array_flip(self::DATATYPE_KEYS));
         foreach ($nonDatatypeKeys as $key => $value) {
             if ($key === 'name') {
                 $columnMetadata[] = [
