--- tests/FunctionalAdapter/basic-data-one-column/config.json
@@ -1,19 +1,0 @@
-{
-  "action": "run",
-  "parameters": {
-    "tables": [
-      {
-        "id": 1,
-        "name": "table1",
-        "table": {
-          "schema": "testdb",
-          "tableName": "table1"
-        },
-        "columns": [
-          "col1"
-        ],
-        "outputTable": "table1"
-      }
-    ]
-  }
-}
