--- tests/FunctionalAdapter/basic-data-query/config.json
@@ -1,13 +1,0 @@
-{
-  "action": "run",
-  "parameters": {
-    "tables": [
-      {
-        "id": 1,
-        "name": "table1",
-        "query": "SELECT * FROM table1",
-        "outputTable": "table1"
-      }
-    ]
-  }
-}
