--- tests/FunctionalAdapter/basic-data-table-incremental-fetch/config.json
@@ -1,13 +1,0 @@
-{
-  "action": "run",
-  "parameters": {
-    "name": "table1",
-    "outputTable": "table1",
-    "table": {
-      "schema": "testdb",
-      "tableName": "table1"
-    },
-    "incrementalFetchingColumn": "id",
-    "incrementalFetchingLimit": 2
-  }
-}
