--- tests/FunctionalAdapter/basic-data-table/config.json
@@ -1,16 +1,0 @@
-{
-  "action": "run",
-  "parameters": {
-    "tables": [
-      {
-        "id": 1,
-        "name": "table1",
-        "table": {
-          "schema": "testdb",
-          "tableName": "table1"
-        },
-        "outputTable": "table1"
-      }
-    ]
-  }
-}
