--- tests/FunctionalAdapter/empty-data/config.json
@@ -1,3 +1,0 @@
-{
-  "action": "run"
-}
